refactoring, bugs fix, adding tests

+++ src/main/java/dao/FilesDao.java
@@ -1,19 +1,12 @@
 package dao;
-
 import models.FileString;
 import org.hibernate.Session;
-import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
 import org.hibernate.query.Query;
 import utils.HibernateSessionFactoryUtil;
-
-
 import java.util.List;
 
-
-
 public class FilesDao {
-
 
     public FileString findById(int id) {
         return HibernateSessionFactoryUtil.getSessionFactory().openSession().get(FileString.class, id);
@@ -40,17 +33,21 @@ public class FilesDao {
         return files;
     }
 
+    public int save(FileString file) {
+        if(file.getPosition()==-100 || file.getFileIndex()==-100 || file.getWriter()==-100){ //we need position for file unless his empty
+            return 0;
+        } else {
+            Session session = HibernateSessionFactoryUtil.getSessionFactory().openSession();
+            Transaction tx1 = session.beginTransaction();
+            session.save(file);
+            tx1.commit();
+            session.close();
+            return 1;
+        }
 
-    public void save(FileString file) {
-        Session session = HibernateSessionFactoryUtil.getSessionFactory().openSession();
-        Transaction tx1 = session.beginTransaction();
-        session.save(file);
-        tx1.commit();
-        session.close();
     }
 
     public void update(FileString file) {
-        System.out.println(file);
         Session session = HibernateSessionFactoryUtil.getSessionFactory().openSession();
         Transaction tx1 = session.beginTransaction();
         session.update(file);
@@ -66,22 +63,24 @@ public class FilesDao {
         session.close();
     }
 
-
-
     public List<FileString> findAll() {
         List<FileString> files = (List<FileString>)  HibernateSessionFactoryUtil.getSessionFactory().openSession().createQuery("FROM FileString ORDER BY position").list();
 
         return files;
     }
 
-        public static void deleteAll() {
+        public static int deleteAll() {
             Session session = HibernateSessionFactoryUtil.getSessionFactory().openSession();
             session.beginTransaction();
             final List<?> instances = session.createCriteria(FileString.class).list();
+            if(instances.size()==0){
+                return 0;
+            }
             for (Object obj : instances) {
                 session.delete(obj);
             }
             session.getTransaction().commit();
+            return 1;
         }
 
 }
